refactor cpp tests/ directory
