+++ foreign/cpp/tests/e2e/message.cpp
@@ -24,7 +24,7 @@
 #include <gtest/gtest.h>
 
 #include "lib.rs.h"
-#include "tests/common/test_helpers.hpp"
+#include "tests/e2e/test_helpers.hpp"
 
 class LowLevelE2E_Message : public E2ETestFixture {};
