+++ foreign/cpp/tests/e2e/stream.cpp
@@ -23,7 +23,7 @@
 #include <gtest/gtest.h>
 
 #include "lib.rs.h"
-#include "tests/common/test_helpers.hpp"
+#include "tests/e2e/test_helpers.hpp"
 
 class LowLevelE2E_Stream : public E2ETestFixture {};
