+++ foreign/cpp/tests/e2e/topic.cpp
@@ -27,7 +27,7 @@
 #include <gtest/gtest.h>
 
 #include "lib.rs.h"
-#include "tests/common/test_helpers.hpp"
+#include "tests/e2e/test_helpers.hpp"
 
 class LowLevelE2E_Topic : public E2ETestFixture {};
