+++ foreign/cpp/tests/unit/message.cpp
